Updated the install_plaso.sh script with new structure
I changed the way the script was grabbing his data and doing
the installation. Now here is the workflow:
1 - Get the latest version of Plaso from the official Github
    repository
2 - Downloading the .tar.gz archive of the latest version from
    the official Github repository
3 - Extracting the archive
4 - Deleting the archive after extraction
5 - Starting the installation

Note that the installation is not working for now, because my
account does not have the rights to write in the folder. I will
have to do a check those rights later on.