Changed the whole behaviour of match_hash.sh
match_hash.sh uses the NSRL Query tools to get a list of the unknown processes
from either a hash file or an entire directory. If it analyses a directory, it
will computes the files hashes before searching them.
The scripts starts a NSRLServer, populates it with the latest database release
and then it with the NSRLLookup tool. It generates an output file in the
hash_outputs/ directory with the name of the folder or the hash file that was
filtered.